Public

TopologicalNumbers.BHZ

TopologicalNumbers.Flux2d

TopologicalNumbers.Haldane

TopologicalNumbers.KaneMele

TopologicalNumbers.KitaevChain

TopologicalNumbers.KitaevHoneycomb

TopologicalNumbers.SSH

TopologicalNumbers.ThoulessPump

TopologicalNumbers.calcBerryFlux

TopologicalNumbers.calcBerryPhase

TopologicalNumbers.calcChern

TopologicalNumbers.calcChernSurface

TopologicalNumbers.calcPhaseDiagram

TopologicalNumbers.calcPhaseDiagram

TopologicalNumbers.calcSecondChern

TopologicalNumbers.calcWeylNode

TopologicalNumbers.calcZ2

TopologicalNumbers.findWeylPoint

TopologicalNumbers.plot1D

TopologicalNumbers.plot1D

TopologicalNumbers.plot1D

TopologicalNumbers.plot2D

TopologicalNumbers.plot2D

TopologicalNumbers.showBand

TopologicalNumbers.BHZ — Method

Hamiltonian of the Bernevig–Hughes–Zhang (BHZ) model.

 BHZ(k::T1, p::T2) where {T1<:Union{AbstractVector,Tuple}, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: two-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the BHZ model is defined as

\[H(k)=\begin{pmatrix} h_{0}(\bm{k})+h_{3}(\bm{k}) & 0 & h_{5}(\bm{k})-ih_{4}(\bm{k}) & 0 \\ 0 & h_{0}(\bm{k})-h_{3}(\bm{k}) & 0 & h_{5}(\bm{k})-ih_{4}(\bm{k}) \\ h_{5}(\bm{k})+ih_{4}(\bm{k}) & 0 & h_{0}(\bm{k})-h_{3}(\bm{k}) & 0 \\ 0 & h_{5}(\bm{k})+ih_{4}(\bm{k}) & 0 & h_{0}(\bm{k})+h_{3}(\bm{k}) \end{pmatrix}\]

\[\begin{align*} h_{0}(\bm{k})=&-(t_{ss}-t_{pp})(\cos(k_{1})+\cos(k_{2}))+\frac{(\epsilon_{s}+\epsilon_{p})}{2} \\ h_{3}(\bm{k})=&2t_{sp}\sin(k_{2}) \\ h_{4}(\bm{k})=&2t_{sp}\sin(k_{1}) \\ h_{5}(\bm{k})=&-(t_{ss}+t_{pp})(\cos(k_{1})+\cos(k_{2}))+\frac{(\epsilon_{s}-\epsilon_{p})}{2} \end{align*}\]

where,,,

TopologicalNumbers.Flux2d — Method

Hamiltonian of the two-dimensional square lattice with flux model.

 Flux2d(k::T1, p::T2) where {T1<:Union{AbstractVector,Tuple}, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: one-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the two-dimensional square lattice with flux model is defined as

\[H(k)=\begin{pmatrix} -2t\cos(k_{2}-2\pi p\frac{1}{q}) & -t & 0 & 0 & \cdots & e^{-ik_{1}} \\ -t & -2t\cos(k_{2}-2\pi p\frac{2}{q}) & -t & 0 & \cdots & 0 \\ 0 & -t & \ddots & -t & \cdots & \vdots \\ \vdots & \vdots & & -t & -2t\cos(k_{2}-2\pi p\frac{(q-1)}{q}) & -t \\ e^{ik_{1}} & 0 & \cdots & 0 & -t & -2t\cos(k_{2}-2\pi p) \end{pmatrix}\]

where,,,

TopologicalNumbers.Haldane — Method

Hamiltonian of the Haldane model.

 Haldane(k::T1, p::T2) where {T1<:Union{AbstractVector,Tuple}, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: two-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the Haldane model is defined as

\[H(\bm{k})=\begin{pmatrix} h_{0}(\bm{k})+h_{3}(\bm{k}) & h_{1}(\bm{k})-ih_{2}(\bm{k}) \\ h_{1}(\bm{k})+ih_{2}(\bm{k}) & h_{0}(\bm{k})-h_{3}(\bm{k}) \end{pmatrix}\]

\[\begin{align*} h_{0}(\bm{k})=&2K\cos(\phi)(\sin(k_{1})+\sin(k_{2})+\sin(k_{1}+k_{2})) \\ h_{1}(\bm{k})=&-J(1+\cos(k_{1})+\cos(k_{2})) \\ h_{2}(\bm{k})=&-J(-\sin(k_{1})+\sin(k_{2})) \\ h_{3}(\bm{k})=&M+2K\sin(\phi)(\sin(k_{1})+\sin(k_{2})-\sin(k_{1}+k_{2})) \end{align*}\]

where,,,

TopologicalNumbers.KaneMele — Method

Hamiltonian of the Kane–Mele model.

 KaneMele(k::T1, p::T2) where {T1<:Union{AbstractVector,Tuple}, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: two-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the Kane–Mele model is defined as

\[H(k)=\begin{pmatrix} h_{3}(\bm{k}) & 0 & h_{5}(\bm{k})-ih_{4}(\bm{k}) & 0 \\ 0 & -h_{3}(\bm{k}) & 0 & h_{5}(\bm{k})-ih_{4}(\bm{k}) \\ h_{5}(\bm{k})+ih_{4}(\bm{k}) & 0 & -h_{3}(\bm{k}) & 0 \\ 0 & h_{5}(\bm{k})+ih_{4}(\bm{k}) & 0 & h_{3}(\bm{k}) \end{pmatrix}\]

\[\begin{align*} h_{3}(\bm{k})=&2\lambda_{\mathrm{SO}}(\sin(k_{1})-\sin(k_{2})-\sin(k_{1}-k_{2})) \\ h_{4}(\bm{k})=&-t(\sin(k_{1})+\sin(k_{2})) \\ h_{5}(\bm{k})=&-t(\cos(k_{1})+\cos(k_{2})+1) \end{align*}\]

where,,,

TopologicalNumbers.KitaevChain — Method

Hamiltonian of the Kitaev chain model.

 KitaevChain(k::T1, p::T2) where {T1<:Real, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: one-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the Kitaev chain model is defined as

\[H(k)=\begin{pmatrix} -\mu-2t\cos(k) & 2i\Delta\sin(k) \\ -2i\Delta\sin(k) & \mu+2t\cos(k) \end{pmatrix}\]

where,,,

TopologicalNumbers.KitaevHoneycomb — Method

Hamiltonian of the Kitaev honeycomb model.

 KitaevHoneycomb(k::T1, p::T2) where {T1<:Union{AbstractVector,Tuple}, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: two-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the Kitaev honeycomb model is defined as

\[H(\bm{k})=\begin{pmatrix} h_{3}(\bm{k}) & h_{1}(\bm{k})-ih_{2}(\bm{k}) \\ h_{1}(\bm{k})+ih_{2}(\bm{k}) & -h_{3}(\bm{k}) \end{pmatrix}\]

\[\begin{align*} h_{1}(\bm{k})=&-\kappa(\sin(k_{2})-\sin(k_{1})+\sin(k_{1}-k_{2})) \\ h_{2}(\bm{k})=&-K(\sin(k_{1})+\sin(k_{2})) \\ h_{3}(\bm{k})=&K(\cos(k_{1})+\cos(k_{2})+1) \end{align*}\]

where,,,

TopologicalNumbers.SSH — Method

Hamiltonian of the Su-Schrieffer-Heeger model.

 SSH(k::T1, p::T2) where {T1<:Real, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: one-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the Su-Schrieffer-Heeger model is defined as

\[H(k)=\begin{pmatrix} 0 & t_{1}+t_{2}e^{-ik} \\ t_{1}+t_{2}e^{ik} & 0 \end{pmatrix}\]

where,,,

TopologicalNumbers.ThoulessPump — Method

Hamiltonian of the Thouless pumping model.

 ThoulessPump(k::T1, p::T2) where {T1<:Union{AbstractVector,Tuple}, T2<:Union{AbstractVector,Tuple}}

Arguments

k::T1: two-dimensional wavenumber k.

p::T2: parameters defined as below.

Definition

Hamiltonian of the Thouless pumping model is defined as

\[H(k)=\]

where,,,

TopologicalNumbers.calcBerryFlux — Method

Calculate the Berry flux in the two-dimensional case with reference to Fukui-Hatsugai-Suzuki method [1].

calcBerryFlux(Hamiltonian::Function, n::Vector{Int64}; N::Int=51, gapless::Real=0.0, rounds::Bool=true)

Arguments

Hamiltionian::Function: the Hamiltonian matrix with one-dimensional wavenumber k as an argument.

n::Vector{Int64}: The wavenumber($2\pi n/N$) when calculating Berry flux.

N::Int=51: The number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

gapless::Real: The threshold that determines the state to be degenerate. Coarsening the mesh(N) but increasing gapless will increase the accuracy of the calculation.

rounds::Bool=true: An option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

Definition

The Berry flux at the wavenumber $\bm{k}$ of the $n$th band $F_{n}(\bm{k})$ is defined by

\[F_{n}(\bm{k})=f_{n}(\bm{k})-df_{n}(\bm{k})\]

\[f_{n}(\bm{k})=\frac{1}{2\pi}\mathrm{Im}\left[\mathrm{Log}\left[U_{n,1}(\bm{k})U_{n,2}(\bm{k}+\bm{e}_{1})U_{n,1}^{*}(\bm{k}+\bm{e}_{2})U_{n,1}^{*}(\bm{k})\right]\right]\]

\[df_{n}(\bm{k})=\frac{1}{2\pi}\mathrm{Im}\left[\mathrm{Log}[U_{n,1}(\bm{k})]+\mathrm{Log}[U_{n,2}(\bm{k}+\bm{e}_{1})]-\mathrm{Log}[U_{n,1}(\bm{k}+\bm{e}_{2})]-\mathrm{Log}[U_{n,1}(\bm{k})]\right]\]

$U_{n,i}(\bm{k})$ is the link variable at wavenumber $\bm{k}$. $\bm{e}_{i}$ is the unit vector.

\[U_{n,i}(\bm{k})=\braket{\Psi_{n}(\bm{k})|\Psi_{n}(\bm{k}+\bm{e}_{i})}\]

$\ket{\Psi_{n}(\bm{k})}$ is the wave function of the $n$th band.

TopologicalNumbers.calcBerryPhase — Method

Calculate the winding numbers in the one-dimensional case.

calcBerryPhase(Hamiltonian::Function; N::Int=51, gapless::Real=0.0, rounds::Bool=true)

Arguments

Hamiltonian::Function: the Hamiltonian matrix function with one-dimensional wavenumber k as an argument.

N::Int: the number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

gapless::Real: the threshold that determines the state to be degenerate. Coarsening the mesh(N) but increasing gapless will increase the accuracy of the calculation.

rounds::Bool: an option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

Definition

The Berry phase of the $n$th band $\nu_{n}$ is defined by

\[\nu_{n}=\frac{1}{\pi}\sum_{k\in\mathrm{BZ}}U_{n}(k)\]

The range $\mathrm{BZ}$(Brillouin Zone) is $k\in[0,2\pi]$. $U_{n,i}(k)$ is the link variable at wavenumber $k$. $e_{1}$ is the unit vector.

\[U_{n}(k)=\braket{\Psi_{n}(k)|\Psi_{n}(k+e_{1})}\]

$\ket{\Psi_{n}(k)}$ is the wave function of the $n$th band.

TopologicalNumbers.calcChern — Method

Calculate the first Chern numbers in the two-dimensional case with reference to Fukui-Hatsugai-Suzuki method [1].

calcChern(Hamiltonian::Function; N::Int=51, gapless::Real=0.0, rounds::Bool=true)

Arguments

Hamiltionian::Function: The Hamiltonian matrix with two-dimensional wavenumber k as an argument.

N::Int=51: The number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

gapless::Real: The threshold that determines the state to be degenerate. Coarsening the mesh(N) but increasing gapless will increase the accuracy of the calculation.

rounds::Bool=true: An option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

Definition

The first Chern number of the $n$th band $\nu_{n}$ is defined by

\[\nu_{n}=\frac{1}{2\pi}\sum_{\bm{k}\in\mathrm{BZ}}\mathrm{Im}\left[\mathrm{Log}\left[U_{n,1}(\bm{k})U_{n,2}(\bm{k}+\bm{e}_{1})U_{n,1}^{*}(\bm{k}+\bm{e}_{2})U_{n,2}^{*}(\bm{k})\right]\right]\]

The range $\mathrm{BZ}$(Brillouin Zone) is $\bm{k}\in[0,2\pi]^{2}$. $U_{n,i}(\bm{k})$ is the link variable at wavenumber $\bm{k}$. $\bm{e}_{i}$ is the unit vector.

\[U_{n,i}(\bm{k})=\braket{\Psi_{n}(\bm{k})|\Psi_{n}(\bm{k}+\bm{e}_{i})}\]

$\ket{\Psi_{n}(\bm{k})}$ is the wave function of the $n$th band.

TopologicalNumbers.calcChernSurface — Method

calcChernSurface(H::Function, kn::String; kn_mesh::Int=51, N::Int=51, gapless::Real=0.0, rounds::Bool=true, plot::Bool=false)

Arguments

Hamiltionian::Function: The Hamiltonian matrix with three-dimensional wavenumber k as an argument.

kn::String: Compute the Chern number of the plane perpendicular to the "kn" direction in Brillouin zone ("k1", "k2", "k3").

kn_mesh::T: Number of mesh in "kn" direction.

N::Int=51: The number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

gapless::Real: The threshold that determines the state to be degenerate. Coarsening the mesh(N) but increasing gapless will increase the accuracy of the calculation.

rounds::Bool=true: An option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

TopologicalNumbers.calcPhaseDiagram — Method

calcPhaseDiagram(H::Function, param_range::T, alg::String; N::Int=51, gapless::Real=0.0, rounds::Bool=true, plot::Bool=false) where {T<:AbstractVector}

TopologicalNumbers.calcPhaseDiagram — Method

calcPhaseDiagram(H::Function, param_range1::T, param_range2::T, alg::String; N::Int=51, gapless::Real=0.0, rounds::Bool=true, plot::Bool=false) where {T<:AbstractVector}

TopologicalNumbers.calcSecondChern — Method

Calculate the second Chern numbers in the four-dimensional case with reference to Fukui-Hatsugai-Suzuki method [1].

calcSecondChern(Hamiltonian::Function; Nfill::T1=nothing, N::T2=(30, 30, 30, 30), returnRealValue::Bool=true) where {T1<:Union{Int,nothing},T2<:Union{AbstractVector,Tuple}}

Arguments

Hamiltionian: The Hamiltonian matrix with two-dimensional wavenumber k as an argument.

Nfill::T1: The filling number. The default value is Hs ÷ 2, where Hs is the size of the Hamiltonian matrix.

N::T2: The numbers of meshes when discretizing the Brillouin Zone. Each element of N is the number of meshes in the x, y, z, and w directions, respectively.

returnRealValue::Bool: An option to return the value of the topological number by an real value. The topological number returns a value of type Float64 when true, and a value of type ComplexF64 when false.

Definition

Examples

TopologicalNumbers.calcWeylNode — Method

Calculate the Weyl node in the three-dimensional case with reference to Fukui-Hatsugai-Suzuki method [1].

calcWeylNode(Hamiltonian::Function, n::Vector{Int64}; N::Int=51, gapless::Real=0.0, rounds::Bool=true)

Arguments

Hamiltionian::Function: the Hamiltonian matrix with three-dimensional wavenumber k as an argument.

n::Vector{Int64}: The wavenumber($2\pi\bm{n}/N$) when calculating Weyl node.

N::Int=51: The number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

gapless::Real: The threshold that determines the state to be degenerate. Coarsening the mesh(N) but increasing gapless will increase the accuracy of the calculation.

rounds::Bool=true: An option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

TopologicalNumbers.calcZ2 — Method

Calculate the $\mathbb{Z}_2$ numbers in the two-dimensional case with reference to Shiozaki method [2].

calcZ2(Hamiltonian::Function; N::Int=50, rounds::Bool=true, TR::Bool=false)

Arguments

Hamiltonian::Function is a matrix with one-dimensional wavenumber k as an argument.

N::Int is the number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

rounds::Bool is an option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

Definition

The $\mathbb{Z}_{2}$ number of the $2n$th (and $2n-1$th) band $\nu_{n}$ is defined by

\[\nu_{n}=F_{n}-\left(P_{n}(0)-P_{n}(\pi)\right)\]

$F_{n}$ is the Berry flux of the $n$th band in the $\mathrm{BZ}'$. The range $\mathrm{BZ}'$ is $\bm{k}\in[0,2\pi]\times[0,\pi]$ half of BZ(Brillouin Zone).

\[F_{n}=\frac{1}{2\pi}\sum_{\bm{k}\in\mathrm{BZ}'}\mathrm{Im}\left[\mathrm{Log}\left[U_{n,1}(\bm{k})U_{n,2}(\bm{k}+\bm{e}_{1})U_{n,1}^{*}(\bm{k}+\bm{e}_{2})U_{n,1}^{*}(\bm{k})\right]\right]\]

$P_{n}(k_{2})$ is the time-reversal polarization at wavenumber $k_{2}$.

\[P_{n}(k_{2})=\frac{1}{2\pi}\frac{\mathrm{Pf}[\omega(0,k_{2})]}{\mathrm{Pf}[\omega(\pi,k_{2})]}\sum_{k_{1}=0}^{\pi-e_{1}}U_{n,1}(\bm{k})\]

$U_{n,i}(\bm{k})$ is the link variable at wavenumber $\bm{k}$. $\bm{e}_{i}$ is the unit vector.

\[U_{n,i}(\bm{k})=\braket{\Psi_{n}(\bm{k})|\Psi_{n}(\bm{k}+\bm{e}_{i})}\]

$\ket{\Psi_{n}(\bm{k})}$ is the wave function of the $2n$th (and $2n-1$th) band. $\omega(\bm{k})$ is the unitary matrix given by

\[\omega(\bm{k})=\bra{\Psi(-\bm{k})}T\ket{\Psi(\bm{k})}\]

$T$ is the time-reversal operator.

TopologicalNumbers.findWeylPoint — Method

findWeylPoint(Hamiltonian::Function; N::Int=10, gapless::T=[1e-1, 1e-2, 1e-3, 1e-4], rounds::Bool=true) where {T<:AbstractVector{Float64}}

Arguments

Hamiltionian::Function: The Hamiltonian matrix with three-dimensional wavenumber k as an argument.

N::Int=10: The number of meshes when discretizing the Brillouin Zone. The $n$th iteration divides the Brillouin zone into $1/N^n$.

gapless<:AbstractVector{Float64}: The threshold that determines the state to be degenerate. The $n$th iteration adopts the threshold value of the $n$th value of the vector. The number of iterations can be varied by the length of the vector.

rounds::Bool=true: An option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

TopologicalNumbers.plot1D — Method

plot1D(result::NamedTuple; labels::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram")

TopologicalNumbers.plot1D — Method

plot1D(nums::T1, param_range::T2; labels::Bool=true, disp::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram") where {T1<:AbstractMatrix,T2<:AbstractVector}

TopologicalNumbers.plot1D — Method

plot1D(nums::T1, param_range::T2; labels::Bool=true, disp::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram") where {T1<:AbstractMatrix,T2<:AbstractVector}

TopologicalNumbers.plot2D — Method

plot2D(result::NamedTuple; labels::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram")

TopologicalNumbers.plot2D — Method

plot2D(nums::T1, param_range1::T2, param_range2::T3; labels::Bool=true, disp::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram") where {T1<:AbstractArray,T2<:AbstractVector,T3<:AbstractVector}

TopologicalNumbers.showBand — Method

Drawing the band structure of the Hamiltonian.

showBand(Hamiltonian::Function; N::Int=51, labels::Bool=true, value::Bool=true, disp::Bool=false, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="Band")

Arguments

Hamiltonian::Function: the Hamiltonian matrix function of wave number $\bm k$.

N::Int: the number of divisions in the wave number space.

labels::Bool: whether to display the labels of the figure.

value::Bool: whether to output the values of the wave number and the energy in the matrix form.

disp::Bool: whether to display the figure.

png::Bool: whether to save the figure as a PNG file.

pdf::Bool: whether to save the figure as a PDF file.

svg::Bool: whether to save the figure as a SVG file.

filename::String: the name of the output file.

\[\]