Public

TopologicalNumbers.calcBerryPhase

TopologicalNumbers.calcChern

TopologicalNumbers.calcPhaseDiagram

TopologicalNumbers.calcPhaseDiagram

TopologicalNumbers.calcZ2

TopologicalNumbers.plot1D

TopologicalNumbers.plot1D

TopologicalNumbers.plot2D

TopologicalNumbers.plot2D

TopologicalNumbers.showBand

TopologicalNumbers.calcBerryPhase — Method

Calculate the winding numbers in the one-dimensional case.

calcBerryPhase(Hamiltonian::Function; N::Int=51, gapless::Real=0.0, rounds::Bool=true)

Arguments

Hamiltonian::Function: the Hamiltonian matrix function with one-dimensional wavenumber k as an argument.

N::Int: the number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

gapless::Real: the threshold that determines the state to be degenerate. Coarsening the mesh(N) but increasing gapless will increase the accuracy of the calculation.

rounds::Bool: an option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

Definition

The Berry phase of the $n$th band $\nu_{n}$ is defined by

\[\nu_{n}=\frac{1}{\pi i}\int_{\mathrm{BZ}}dkA_{n}(k)\]

The integral range $\mathrm{BZ}$(Brillouin Zone) is $k\in[0,2\pi]$. $A_{n}(k)$ is the Berry conection at wavenumber $k$.

\[A_{n}(k)=\bra{\Psi_{n}(k)}\partial_{k}\ket{\Psi_{n}(k)}\]

$\ket{\Psi_{n}(k)}$ is the wave function of the $n$th band.

TopologicalNumbers.calcChern — Method

Calculate the first Chern numbers in the two-dimensional case with reference to Fukui-Hatsugai-Suzuki method [1].

calcChern(Hamiltonian::Function; N::Int=51, gapless::Real=0.0, rounds::Bool=true)

Arguments

Hamiltionian::Function: the Hamiltonian matrix with one-dimensional wavenumber k as an argument.

N::Int=51: The number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

gapless::Real: The threshold that determines the state to be degenerate. Coarsening the mesh(N) but increasing gapless will increase the accuracy of the calculation.

rounds::Bool=true: An option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

Definition

The firs Chern number of the $n$th band $\nu_{n}$ is defined by

\[\nu_{n}=\frac{1}{2\pi i}\int_{\mathrm{BZ}}d\bm{k}\left(\partial_{k_{1}}A_{n,2}(\bm{k})-\partial_{k_{2}}A_{n,1}(\bm{k})\right)\]

The integral range $\mathrm{BZ}$(Brillouin Zone) is $\bm{k}\in[0,2\pi]^{2}$. $A_{n,i}(\bm{k})$ is the Berry connection at wavenumber $\bm{k}$.

\[A_{n,i}(\bm{k})=\bra{\Psi_{n}(\bm{k})}\partial_{k_{i}}\ket{\Psi_{n}(\bm{k})}\]

$\ket{\Psi_{n}(\bm{k})}$ is the wave function of the $n$th band.

TopologicalNumbers.calcPhaseDiagram — Method

calcPhaseDiagram(H::Function, param_range::T, alg::String; N::Int=51, gapless::Real=0.0, rounds::Bool=true, plot::Bool=false) where {T<:AbstractVector}

TopologicalNumbers.calcPhaseDiagram — Method

calcPhaseDiagram(H::Function, param_range1::T, param_range2::T, alg::String; N::Int=51, gapless::Real=0.0, rounds::Bool=true, plot::Bool=false) where {T<:AbstractVector}

TopologicalNumbers.calcZ2 — Method

Calculate the $\mathbb{Z}_2$ numbers in the two-dimensional case with reference to Shiozaki method [2].

calcZ2(Hamiltonian::Function; N::Int=50, rounds::Bool=true, TR::Bool=false)

Arguments

Hamiltonian::Function is a matrix with one-dimensional wavenumber k as an argument.

N::Int is the number of meshes when discretizing the Brillouin Zone. It is preferable for N to be an odd number to increase the accuracy of the calculation.

rounds::Bool is an option to round the value of the topological number to an integer value. The topological number returns a value of type Int when true, and a value of type Float when false.

Definition

The $\mathbb{Z}_{2}$ number of the $2n$th (and $2n-1$th) band $\nu_{n}$ is defined by

\[\nu_{n}=\frac{1}{2\pi i}\int_{\mathrm{BZ}'}d\bm{k}\left(\partial_{k_{1}}A_{n,2}(\bm{k})-\partial_{k_{2}}A_{n,1}(\bm{k})\right)-\frac{1}{2\pi i}\oint_{\partial\mathrm{BZ}'}dk_{i}A_{n,j\neq i}(\bm{k})\]

The integral range $\mathrm{BZ}'$ is $\bm{k}\in[0,2\pi]\times[0,\pi]$ half of BZ(Brillouin Zone). $A_{n,i}(\bm{k})$ is the Berry connection at wavenumber $\bm{k}$.

\[A_{n,i}(\bm{k})=\bra{\Psi_{n}(\bm{k})}\partial_{k_{i}}\ket{\Psi_{n}(\bm{k})}\]

$\ket{\Psi_{n}(\bm{k})}$ is the wave function of the $2n$th (and $2n-1$th) band.

TopologicalNumbers.plot1D — Method

plot1D(result::NamedTuple; labels::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram")

TopologicalNumbers.plot1D — Method

plot1D(nums::Matrix, param_range::T; labels::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram") where {T<:AbstractVector}

TopologicalNumbers.plot2D — Method

plot2D(result::NamedTuple; labels::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram")

TopologicalNumbers.plot2D — Method

plot2D(nums::T1, param_range1::T2, param_range2::T2; labels::Bool=true, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="phaseDiagram") where {T1<:AbstractArray,T2<:AbstractVector}

TopologicalNumbers.showBand — Method

Drawing the band structure of the Hamiltonian.

showBand(Hamiltonian::Function; N::Int=51, labels::Bool=true, value::Bool=true, disp::Bool=false, png::Bool=false, pdf::Bool=false, svg::Bool=false, filename::String="Band")

Arguments

Hamiltonian::Function: the Hamiltonian matrix function of wave number $\bm k$.

N::Int: the number of divisions in the wave number space.

labels::Bool: whether to display the labels of the figure.

value::Bool: whether to output the values of the wave number and the energy in the matrix form.

disp::Bool: whether to display the figure.

png::Bool: whether to save the figure as a PNG file.

pdf::Bool: whether to save the figure as a PDF file.

svg::Bool: whether to save the figure as a SVG file.

filename::String: the name of the output file.

\[\]